+ a few more changes to headers: simplifying nfs interface and putting tiny/micro-ecm into factor common
