+++ build.vc19/nfs/nfs.vcxproj.filters
@@ -56,5 +56,8 @@
     <ClInclude Include="..\..\include\msieve_common.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\include\nfs_impl.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
