+++ factor/nfs/nfs_filemanip.c
@@ -12,7 +12,7 @@ benefit from your work.
        				   --bbuhrow@gmail.com 12/6/2012
 ----------------------------------------------------------------------*/
 
-#include "nfs.h"
+#include "nfs_impl.h"
 #include "gmp_xface.h"
 
 #ifdef USE_NFS
